+++ src/domain/imap/imap_service.rs
@@ -8,20 +8,21 @@ use native_tls::{TlsConnector, TlsStream};
 use std::{
     collections::HashSet,
     convert::{TryFrom, TryInto},
-    iter::FromIterator,
     net::TcpStream,
+    thread,
 };
 
 use crate::{
     config::{Account, Config},
     domain::{Envelope, Envelopes, Flags, Mbox, Mboxes, Msg, RawEnvelopes, RawMboxes},
+    output::run_cmd,
 };
 
 type ImapSession = imap::Session<TlsStream<TcpStream>>;
 
 pub trait ImapServiceInterface<'a> {
     fn notify(&mut self, config: &Config, keepalive: u64) -> Result<()>;
-    fn watch(&mut self, keepalive: u64) -> Result<()>;
+    fn watch(&mut self, account: &Account, keepalive: u64) -> Result<()>;
     fn fetch_mboxes(&'a mut self) -> Result<Mboxes>;
     fn fetch_envelopes(&mut self, page_size: &usize, page: &usize) -> Result<Envelopes>;
     fn fetch_envelopes_with(
@@ -58,7 +59,7 @@ pub struct ImapService<'a> {
 
 impl<'a> ImapService<'a> {
     fn sess(&mut self) -> Result<&mut ImapSession> {
-        if let None = self.sess {
+        if self.sess.is_none() {
             debug!("create TLS builder");
             debug!("insecure: {}", self.account.imap_insecure);
             let builder = TlsConnector::builder()
@@ -122,12 +123,17 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
     }
 
     fn fetch_envelopes(&mut self, page_size: &usize, page: &usize) -> Result<Envelopes> {
+        debug!("fetch envelopes");
+        debug!("page size: {:?}", page_size);
+        debug!("page: {:?}", page);
+
         let mbox = self.mbox.to_owned();
         let last_seq = self
             .sess()?
             .select(&mbox.name)
             .context(format!(r#"cannot select mailbox "{}""#, self.mbox.name))?
             .exists as i64;
+        debug!("last sequence number: {:?}", last_seq);
 
         if last_seq == 0 {
             return Ok(Envelopes::default());
@@ -142,13 +148,14 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
         } else {
             String::from("1:*")
         };
+        debug!("range: {:?}", range);
 
         let fetches = self
             .sess()?
-            .fetch(range, "(ENVELOPE FLAGS INTERNALDATE)")
-            .context(r#"cannot fetch messages within range "{}""#)?;
+            .fetch(&range, "(ENVELOPE FLAGS INTERNALDATE)")
+            .context(format!(r#"cannot fetch messages within range "{}""#, range))?;
         self._raw_msgs_cache = Some(fetches);
-        Ok(Envelopes::try_from(self._raw_msgs_cache.as_ref().unwrap())?)
+        Envelopes::try_from(self._raw_msgs_cache.as_ref().unwrap())
     }
 
     fn fetch_envelopes_with(
@@ -186,7 +193,7 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
             .fetch(&range, "(ENVELOPE FLAGS INTERNALDATE)")
             .context(r#"cannot fetch messages within range "{}""#)?;
         self._raw_msgs_cache = Some(fetches);
-        Ok(Envelopes::try_from(self._raw_msgs_cache.as_ref().unwrap())?)
+        Envelopes::try_from(self._raw_msgs_cache.as_ref().unwrap())
     }
 
     /// Find a message by sequence number.
@@ -201,9 +208,9 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
             .context(r#"cannot fetch messages "{}""#)?;
         let fetch = fetches
             .first()
-            .ok_or(anyhow!(r#"cannot find message "{}"#, seq))?;
+            .ok_or_else(|| anyhow!(r#"cannot find message "{}"#, seq))?;
 
-        Ok(Msg::try_from(fetch)?)
+        Msg::try_from(fetch)
     }
 
     fn find_raw_msg(&mut self, seq: &str) -> Result<Vec<u8>> {
@@ -217,14 +224,14 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
             .context(r#"cannot fetch raw messages "{}""#)?;
         let fetch = fetches
             .first()
-            .ok_or(anyhow!(r#"cannot find raw message "{}"#, seq))?;
+            .ok_or_else(|| anyhow!(r#"cannot find raw message "{}"#, seq))?;
 
         Ok(fetch.body().map(Vec::from).unwrap_or_default())
     }
 
     fn append_raw_msg_with_flags(&mut self, mbox: &Mbox, msg: &[u8], flags: Flags) -> Result<()> {
         self.sess()?
-            .append(&mbox.name, &msg)
+            .append(&mbox.name, msg)
             .flags(flags.0)
             .finish()
             .context(format!(r#"cannot append message to "{}""#, mbox.name))?;
@@ -242,16 +249,21 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
     }
 
     fn notify(&mut self, config: &Config, keepalive: u64) -> Result<()> {
+        debug!("notify");
+
         let mbox = self.mbox.to_owned();
 
-        debug!("examine mailbox: {}", mbox.name);
+        debug!("examine mailbox {:?}", mbox);
         self.sess()?
             .examine(&mbox.name)
-            .context(format!("cannot examine mailbox `{}`", &self.mbox.name))?;
+            .context(format!("cannot examine mailbox {}", self.mbox.name))?;
 
         debug!("init messages hashset");
-        let mut msgs_set: HashSet<u32> =
-            HashSet::from_iter(self.search_new_msgs()?.iter().cloned());
+        let mut msgs_set: HashSet<u32> = self
+            .search_new_msgs()?
+            .iter()
+            .cloned()
+            .collect::<HashSet<_>>();
         trace!("messages hashset: {:?}", msgs_set);
 
         loop {
@@ -271,7 +283,7 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
             let uids: Vec<u32> = self
                 .search_new_msgs()?
                 .into_iter()
-                .filter(|uid| msgs_set.get(&uid).is_none())
+                .filter(|uid| -> bool { msgs_set.get(uid).is_none() })
                 .collect();
             debug!("found {} new messages not in hashset", uids.len());
             trace!("messages hashet: {:?}", msgs_set);
@@ -309,7 +321,7 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
         }
     }
 
-    fn watch(&mut self, keepalive: u64) -> Result<()> {
+    fn watch(&mut self, account: &Account, keepalive: u64) -> Result<()> {
         debug!("examine mailbox: {}", &self.mbox.name);
         let mbox = self.mbox.to_owned();
 
@@ -330,8 +342,17 @@ impl<'a> ImapServiceInterface<'a> for ImapService<'a> {
                     })
                 })
                 .context("cannot start the idle mode")?;
-            // FIXME
-            // ctx.config.exec_watch_cmds(&ctx.account)?;
+
+            let cmds = account.watch_cmds.clone();
+            thread::spawn(move || {
+                debug!("batch execution of {} cmd(s)", cmds.len());
+                cmds.iter().for_each(|cmd| {
+                    debug!("running command {:?}…", cmd);
+                    let res = run_cmd(cmd);
+                    debug!("{:?}", res);
+                })
+            });
+
             debug!("end loop");
         }
     }